rename refactored legacyBuildWords to getWordFreqLookupNLP
